Terminate tests with SIGALRM if they take more than 120 seconds
The Debian autobuilders run cogl's build-time tests to get an idea of
how much can be expected to work on each supported architecture.
We're currently ignoring test failures and running the tests for
information only, but if a test hangs, that will still cause the build
to fail (and tie up an autobuilder until a rather long timeout is
reached).

The fork of cogl in Mutter uses a 120 second timeout for tests in its
Meson build system, but other than that it's purely an arbitrary number.
dbus uses a similar alarm()-based watchdog, as do a couple of tests
in GLib.

Bug-Debian: https://bugs.debian.org/935592
Signed-off-by: Simon McVittie <smcv@debian.org>
Forwarded: https://gitlab.gnome.org/GNOME/cogl/merge_requests/11

Gbp-Pq: Name Terminate-tests-with-SIGALRM-if-they-take-more-than-120-s.patch