Terminate tests with SIGALRM if they take more than 120 seconds
We're currently ignoring test failures, but if a test hangs, that will
still cause the build to fail (and tie up a buildd for 150 minutes).

The fork of cogl in Mutter uses a 120 second timeout for tests in its
Meson build system, but other than that it's purely an arbitrary number.

Bug-Debian: https://bugs.debian.org/935592
Forwarded: no

Gbp-Pq: Name Terminate-tests-with-SIGALRM-if-they-take-more-than-120-s.patch